refactor(hindsight): one netting module
netting_decoders.rs and intents/netting.rs merge into netting.rs: the
engine (sender_flow, venue_flow, find_intent_trade) plus the two
generic decoders (SenderNetting, IntentNetting) in one place. Venue
netting decoders stay in venues/ and call the engine. Decoder name
labels in the records are unchanged.
