+++ tools/hindsight/src/decoder/venues/rabby.rs
@@ -18,7 +18,7 @@ use async_trait::async_trait;
 
 use crate::decoder::{
     decode::{DecodeContext, TradeDecoder, TraderFlow},
-    netting_decoders::venue_flow,
+    netting::venue_flow,
     registry::VenueAddresses,
 };
